requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.1
django-ckeditor-5==0.2.15
django-ninja==1.3.0
djangorestframework==3.15.2
docker==7.1.0
idna==3.10
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
